cloud_test_app/requirements.txt
Django>=3.2.6
psycopg2>=2.9.1
mysqlclient>=2.0.3
django-db-locking>=2.0.4
requests>=2.26.0
tenacity>=8.0.1
django-bootstrap4>=3.0.1
